Tuning and refactor code to load all ngram datasets
