+++ airflow/dags/ngrametl_airflow/build_load_ngram_dag.py
@@ -1,6 +1,5 @@
 from datetime import datetime, timedelta
-from airflow import models, configuration
-from airflow.operators.python import PythonOperator
+from airflow import models
 from airflow.providers.google.cloud.operators.dataflow import (
     DataflowStartFlexTemplateOperator,
 )
@@ -8,13 +7,14 @@ import os
 
 from google.cloud import bigquery
 
-from wordetl_airflow.bigquery_utils import submit_bigquery_job
-from wordetl_airflow.file_utils import read_file
+from ngrametl_airflow.bigquery_utils import submit_bigquery_job
+from ngrametl_airflow.file_utils import read_file
 
 
 def build_load_ngram_dag(
     dag_id,
-    n_gram,
+    dataset,
+    ngram,
     input_file,
     output_table,
     dataset_project_id,
@@ -23,14 +23,12 @@ def build_load_ngram_dag(
     dataflow_environment,
     notification_emails=None,
     load_start_date=datetime(2021, 1, 1),
-    load_end_date=None,
-    load_schedule_interval=None,
     load_max_active_runs=1,
 ):
     default_dag_args = {
         "depends_on_past": False,
         "start_date": load_start_date,
-        "end_date": load_end_date,
+        "end_date": None,
         "email_on_failure": True,
         "email_on_retry": False,
         "retries": 0,
@@ -44,22 +42,23 @@ def build_load_ngram_dag(
     dag = models.DAG(
         dag_id,
         catchup=False,
-        schedule_interval=load_schedule_interval,
+        schedule_interval=None,
         max_active_runs=load_max_active_runs,
         default_args=default_dag_args,
     )
 
-    table_name = output_table.format(n=n_gram)
-
     dataflow_operator = DataflowStartFlexTemplateOperator(
-        task_id="start_flex_template_streaming_beam_sql",
+        task_id="load_ngram",
         body={
             "launchParameter": {
                 "containerSpecGcsPath": f"{dataflow_template_path}/ngrams-beam.json",
-                "jobName": 'ngram-beam-{{ execution_date.strftime("%Y%m%d-%H%M%S") }}',
+                "jobName": "load-{dataset}-{ngram}".format(
+                    dataset=dataset.replace("_", "-"), ngram=ngram
+                )
+                + '-{{ execution_date.strftime("%Y%m%d-%H%M%S") }}',
                 "parameters": {
-                    "input-file": input_file.format(n=n_gram),
-                    "output-table": f"{dataset_project_id}:{dataset_name}.{table_name}",
+                    "input-file": input_file,
+                    "output-table": f"{dataset_project_id}:{dataset_name}.{output_table}",
                 },
                 "environment": dataflow_environment,
             }
@@ -80,7 +79,7 @@ def build_load_ngram_dag(
         sql_template = read_file(sql_path)
 
         template_context = {
-            "table": table_name,
+            "table": output_table,
             "project_id": dataset_project_id,
             "dataset": dataset_name,
         }
@@ -89,16 +88,5 @@ def build_load_ngram_dag(
         job = client.query(sql, job_config=job_config, timeout=7200)
         submit_bigquery_job(job, job_config)
         assert job.state == "DONE"
-
-    enrich_operator = PythonOperator(
-        task_id="enrich",
-        python_callable=enrich_task,
-        provide_context=True,
-        execution_timeout=timedelta(minutes=130),
-        retries=0,
-        dag=dag,
-    )
-
-    dataflow_operator >> enrich_operator
 
     return dag
